Actualizo tipos a float en puntos, migración init_db con NUMERIC(10,2)

+++ api/schemas.py
@@ -4,5 +4,5 @@ class PremiosResponse(BaseModel):
     dni: str = Field(..., example="47178389")
     canasta: int = Field(..., example=146)
     pavo: int = Field(..., example=168)
-    puntos: int = Field(0, example=0)
+    puntos: float = Field(0.0, example=0.0)
     pv: str = Field("", example="PV47178389")
